+++ v3/newrelic/internal_test.go
@@ -1401,8 +1401,15 @@ func TestExternalSegmentCustomFieldsWithURL(t *testing.T) {
 				"category":         "generic",
 				"nr.entryPoint":    true,
 			},
-			UserAttributes:  map[string]interface{}{},
-			AgentAttributes: map[string]interface{}{},
+			UserAttributes: map[string]interface{}{},
+			AgentAttributes: map[string]interface{}{
+				"request.uri":                   "/hello",
+				"request.headers.contentType":   "text/html; charset=utf-8",
+				"request.headers.host":          "my_domain.com",
+				"request.method":                "GET",
+				"request.headers.contentLength": 753,
+				"request.headers.accept":        "text/plain",
+			},
 		},
 	})
 }
@@ -1458,8 +1465,15 @@ func TestExternalSegmentCustomFieldsWithRequest(t *testing.T) {
 				"category":         "generic",
 				"nr.entryPoint":    true,
 			},
-			UserAttributes:  map[string]interface{}{},
-			AgentAttributes: map[string]interface{}{},
+			UserAttributes: map[string]interface{}{},
+			AgentAttributes: map[string]interface{}{
+				"request.uri":                   "/hello",
+				"request.headers.contentType":   "text/html; charset=utf-8",
+				"request.headers.host":          "my_domain.com",
+				"request.method":                "GET",
+				"request.headers.contentLength": 753,
+				"request.headers.accept":        "text/plain",
+			},
 		},
 	})
 }
@@ -1523,8 +1537,15 @@ func TestExternalSegmentCustomFieldsWithResponse(t *testing.T) {
 				"category":         "generic",
 				"nr.entryPoint":    true,
 			},
-			UserAttributes:  map[string]interface{}{},
-			AgentAttributes: map[string]interface{}{},
+			UserAttributes: map[string]interface{}{},
+			AgentAttributes: map[string]interface{}{
+				"request.uri":                   "/hello",
+				"request.headers.contentType":   "text/html; charset=utf-8",
+				"request.headers.host":          "my_domain.com",
+				"request.method":                "GET",
+				"request.headers.contentLength": 753,
+				"request.headers.accept":        "text/plain",
+			},
 		},
 	})
 }